Time: 1 ms (86.61%), Space: 41.1 MB (39.4%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -5,11 +5,11 @@ class Solution {
         int ans = low + ((high-low) / 2);
         while (low <= high) {
             int mid = low + ((high-low) / 2);
-            if (1l*mid * mid  > x) {
-                high = (int) mid - 1;
+            if ( 1l * mid * mid  > x) {
+                high =  mid - 1;
             } else {
-                ans = (int) mid;
-                low = (int) mid + 1;
+                ans =  mid;
+                low =  mid + 1;
             }
         }
         return ans;
